Refactor stars cross-matching benchmark
